some refactoring for better testability
